requirements.txt
numpy==1.17.4
opencv-python==4.2.0.32
tensorflow==2.7.0
python-vlc==3.0.7110

